refactor(cli): split brief.ts into _brief_render and _brief_assembly to respect 500-line limit
